index/mapb_write: Add an inverted index encoder
Assuming the inverted index is a mapping of words to a list of Postings, this
takes the index and writes a sequence of bytes representing the object.

To enable fast retrieval, the file uses a key-value pair consisting of data
chunk offsets and sequence lengths. The offsets are used to reference a variable
length sequence of bytes. The reason for doing this allows a reader to
deterministically skip over certain key-value pairs since all pairs have a
consistent size.

The data chunk itself is a sequence of variable length byte sequences, which
makes it difficult to efficiently traverse / skip over the sequences. Offsets
from the key-value pairs within the ref chunk help solve this problem.