index/writer: Add new writer to handle partial indices.
The previous index encoder could not handle partial writes, so the entire
inverted index had to be in-memory before it could be stored.

This partitions the writing process so that a file can store multiple partial
indices at once.

We defer efficient traversal at a later step in the merging process, which
remains incomplete.

To document the file formats, a new file called `index/spec` is introduced.