index/writer: Add index merging function
This allows the ability to merge a partial index into a complete index. To merge
the indices, it performs a k-way merge on all the partial indices in the file
using a priority queue on the keys and values.

The spec has been updated to accomodate this new `merged` format.