indexlib/writer: Make separate index directory
Each token is stored in its own bucket within the index directory, which
corresponds to its first grapheme. If it's ASCII, the bucket corresponds to its
decimal value. Otherwise, it's stored in the `misc` bucket.

Seek files have been added for each bucket, which contains a sequence of sorted
key-u32 pairs.

The bucket itself has been modified to only include a list of Postings in
sequence without the string keys.