lib/reader: Expose linked docids, remove docID getter from URL
It's unfriendly for the user to rely on retrieving URLs from docids since that
would mean the user would have to independently parse the doclinks file.

Therefore, we remove the method in favor of doclinks initialization (separate)
and retreiving linked docids from the docid.

This also reverts previous changes to search/searcht which affected the
parameters of base initialization.