lib/queryproc: Improve query processor and params
"lib/queryproc" exists to replace "lib/process_query".

There were several issues with the last processor:

    -   the query engine did not use lib.tokenize,
        which greatly increased the chance of token misses,
        and did not account for uppercase queries.

    -   the score computations forgo TF-IDF sums for cosine
        similarity, which means that a document with less value
        containing the query could match equally with one
        with more value; therefore, it's necessary to include
        TF-IDF sums (which is a basic requirement!)

    -   no tunable parameters; this is addressed with lib/params

    -   scores were not normalized properly for the static
        quality score, which led to several scores being the same
        for multiple queries and overpowering the relevance scores

    -   the overall code is rather messy, and the query processing
        was not independent of the interface (web gui, terminal);
        therefore, process_query no longer returns any formatted
        results

    -   queries did not handle stopwords, which could've been abused to
        inflate query times

This also addresses the lack of modularity of the last processor. That is
"search.py" and "searcht.py" had independent copies of the query processor,
which is not ideal when making changes to how queries are processed.

These changes can be seen in "search.py" and "searcht.py".
Note: This does not remove old code in case changes need to be made.