makeindex: Distinguish important tags
Important tags are no longer treated equally. Instead, a distinct value shall be
assigned to each tag with the first tags being the most important. This allows
for fine tuning of the importance factors as seen in lib/params.

The spec is updated to not mention struct posting's bitfield field as an array
of booleans even though it's a reasonable interpretation.

To accomodate this change, Posting now reads the important field as 4 bits,
which allows up to 15 other distinct tags excluding no tag.