lib/queryproc: Improve query pre-processing
This eliminates tokens that do not have any documents before it actually
processes the query. This also adjusts how tokens are counted when determining
the percentage of stopwords in the query.

However, if a substantial amount of non-unique tokens were pruned, the query is
discarded since it probably wasn't useful in the first place.

Additionally, an LRU cache is added which should speed up retrieval for postings
on subsequent reads (reading from disk is SLOOOOOW). Tokens are read in sorted
order to make it more cache friendly in case it reads tokens from the same
bucket in sequence.